fix minor

+++ Services/Interfaces/IUserAddressService.cs
@@ -9,6 +9,6 @@ namespace Services.Interfaces
 {
     public interface IUserAddressService : IBaseService<UserAddressDetailDto, UserAddressCreateDto, UserAddressUpdateDto>
     {
-        public Task<bool> isValidVietnamAddress(int longtitude, int latitude);
+        public Task<bool> isValidVietnamAddress(decimal longtitude, decimal latitude);
     }
 }
